+++ Source/IGListBatchContext.h
@@ -66,7 +66,37 @@ NS_SWIFT_NAME(ListBatchContext)
  @param sectionController The section controller who's cells need reloading.
  */
 - (void)reloadSectionController:(IGListSectionController *)sectionController;
+    
+/**
+ Moves a section controller from one index to another during interactive reordering.
+ 
+ @param sectionController The section controller to move.
+ @param fromIndex The index where the section currently resides.
+ @param toIndex The index the section should move to.
+ */
+- (void)moveSectionControllerInteractive:(IGListSectionController *)sectionController
+                               fromIndex:(NSInteger)fromIndex
+                                 toIndex:(NSInteger)toIndex NS_AVAILABLE_IOS(9_0);
 
+/**
+ Moves an object within a section controller from one index to another during interactive reordering.
+ 
+ @param sectionController The section controller containing the object to move.
+ @param fromIndex The index where the object currently resides.
+ @param toIndex The index the object should move to.
+ */
+- (void)moveInSectionControllerInteractive:(IGListSectionController *)sectionController
+                                 fromIndex:(NSInteger)fromIndex
+                                   toIndex:(NSInteger)toIndex NS_AVAILABLE_IOS(9_0);
+    
+/**
+ Reverts an move from one indexPath to another during interactive reordering.
+ 
+ @param sourceIndexPath The indexPath the item was originally in.
+ @param destinationIndexPath The indexPath the item was moving to.
+ */
+- (void)revertInvalidInteractiveMoveFromIndexPath:(NSIndexPath *)sourceIndexPath
+                                      toIndexPath:(NSIndexPath *)destinationIndexPath NS_AVAILABLE_IOS(9_0);
 @end
 
 NS_ASSUME_NONNULL_END